# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ABZ Market API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: string;
  name: string;
  price: number;
  oldPrice?: number;
  description?: string;
  images?: string[];
  categoryId: string;
  categoryName?: string;
  storeId: string;
  storeName: string;
  rating?: number;
  reviewCount?: number;
  colors?: string[];
  sizes?: string[];
  dimensions?: string;
  warranty?: string;
  deliveryDays?: number;
  isTopSelling?: boolean;
  isFeatured?: boolean;
  discount?: number;
  salesCount?: number;
}

export interface Category {
  id: string;
  name: string;
  icon?: string;
  productCount?: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  new: "new",
  accepted: "accepted",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export type OrderPaymentMethod =
  (typeof OrderPaymentMethod)[keyof typeof OrderPaymentMethod];

export const OrderPaymentMethod = {
  cash: "cash",
  card: "card",
  installment: "installment",
} as const;

export interface OrderItem {
  productId: string;
  productName: string;
  productImage?: string;
  quantity: number;
  price: number;
  color?: string;
}

export interface Order {
  id: string;
  orderNumber: string;
  status: OrderStatus;
  items: OrderItem[];
  customerName?: string;
  customerPhone?: string;
  address?: string;
  comment?: string;
  paymentMethod?: OrderPaymentMethod;
  totalPrice: number;
  createdAt: string;
  storeId?: string;
  storeName?: string;
}

export type CreateOrderRequestItemsItem = {
  productId: string;
  quantity: number;
  color?: string;
};

export type CreateOrderRequestPaymentMethod =
  (typeof CreateOrderRequestPaymentMethod)[keyof typeof CreateOrderRequestPaymentMethod];

export const CreateOrderRequestPaymentMethod = {
  cash: "cash",
  card: "card",
  installment: "installment",
} as const;

export interface CreateOrderRequest {
  items: CreateOrderRequestItemsItem[];
  customerName: string;
  customerPhone: string;
  address: string;
  comment?: string;
  paymentMethod: CreateOrderRequestPaymentMethod;
}

export type StoreType = (typeof StoreType)[keyof typeof StoreType];

export const StoreType = {
  official: "official",
  manufacturer: "manufacturer",
  partner: "partner",
} as const;

export interface Store {
  id: string;
  name: string;
  logo?: string;
  coverImage?: string;
  description?: string;
  type: StoreType;
  rating?: number;
  reviewCount?: number;
  productCount?: number;
  salesCount?: number;
  deliveryRate?: number;
  location?: string;
  phone?: string;
  stir?: string;
  activityType?: string;
  isVerified?: boolean;
}

export interface CreateStoreRequest {
  name: string;
  activityType: string;
  stir?: string;
  location: string;
  phone: string;
  description?: string;
}

export type AnalyticsSummaryTopProductsItem = {
  productId: string;
  productName: string;
  revenue: number;
  salesCount: number;
};

export interface AnalyticsSummary {
  totalRevenue: number;
  revenueChange: number;
  totalOrders: number;
  ordersChange: number;
  averageCheck: number;
  averageCheckChange: number;
  conversionRate: number;
  conversionChange: number;
  activeProducts: number;
  topProducts: AnalyticsSummaryTopProductsItem[];
}

export interface ChartDataPoint {
  date: string;
  revenue: number;
  orders: number;
}

export type GetProductsParams = {
  categoryId?: string;
  search?: string;
  minPrice?: number;
  maxPrice?: number;
  color?: string;
  size?: string;
  storeId?: string;
  featured?: boolean;
  limit?: number;
  offset?: number;
};

export type GetProducts200 = {
  products: Product[];
  total: number;
};

export type GetCategories200 = {
  categories: Category[];
};

export type GetOrdersParams = {
  status?: GetOrdersStatus;
};

export type GetOrdersStatus =
  (typeof GetOrdersStatus)[keyof typeof GetOrdersStatus];

export const GetOrdersStatus = {
  new: "new",
  accepted: "accepted",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export type GetOrders200 = {
  orders: Order[];
};

export type UpdateOrderStatusBodyStatus =
  (typeof UpdateOrderStatusBodyStatus)[keyof typeof UpdateOrderStatusBodyStatus];

export const UpdateOrderStatusBodyStatus = {
  new: "new",
  accepted: "accepted",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export type UpdateOrderStatusBody = {
  status: UpdateOrderStatusBodyStatus;
};

export type GetStoresParams = {
  search?: string;
  type?: GetStoresType;
};

export type GetStoresType = (typeof GetStoresType)[keyof typeof GetStoresType];

export const GetStoresType = {
  official: "official",
  manufacturer: "manufacturer",
  partner: "partner",
} as const;

export type GetStores200 = {
  stores: Store[];
};

export type GetAnalyticsSummaryParams = {
  period?: GetAnalyticsSummaryPeriod;
};

export type GetAnalyticsSummaryPeriod =
  (typeof GetAnalyticsSummaryPeriod)[keyof typeof GetAnalyticsSummaryPeriod];

export const GetAnalyticsSummaryPeriod = {
  today: "today",
  "7days": "7days",
  "30days": "30days",
  "90days": "90days",
} as const;

export type GetAnalyticsChartParams = {
  period?: GetAnalyticsChartPeriod;
};

export type GetAnalyticsChartPeriod =
  (typeof GetAnalyticsChartPeriod)[keyof typeof GetAnalyticsChartPeriod];

export const GetAnalyticsChartPeriod = {
  "7days": "7days",
  "30days": "30days",
  "90days": "90days",
} as const;

export type GetAnalyticsChart200 = {
  data: ChartDataPoint[];
};