# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ABZ Market API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product operations
  - name: categories
    description: Category operations
  - name: orders
    description: Order operations
  - name: stores
    description: Store operations
  - name: cart
    description: Cart operations
  - name: analytics
    description: Analytics operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: Get products list
      parameters:
        - name: categoryId
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: color
          in: query
          schema:
            type: string
        - name: size
          in: query
          schema:
            type: string
        - name: storeId
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      $ref: "#/components/schemas/Product"
                  total:
                    type: integer
                required:
                  - products
                  - total

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /categories:
    get:
      operationId: getCategories
      tags: [categories]
      summary: Get all categories
      responses:
        "200":
          description: Categories list
          content:
            application/json:
              schema:
                type: object
                properties:
                  categories:
                    type: array
                    items:
                      $ref: "#/components/schemas/Category"
                required:
                  - categories

  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: Get orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [new, accepted, shipped, delivered, cancelled]
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      $ref: "#/components/schemas/Order"
                required:
                  - orders
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [new, accepted, shipped, delivered, cancelled]
              required:
                - status
      responses:
        "200":
          description: Order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /stores:
    get:
      operationId: getStores
      tags: [stores]
      summary: Get stores list
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [official, manufacturer, partner]
      responses:
        "200":
          description: Stores list
          content:
            application/json:
              schema:
                type: object
                properties:
                  stores:
                    type: array
                    items:
                      $ref: "#/components/schemas/Store"
                required:
                  - stores
    post:
      operationId: createStore
      tags: [stores]
      summary: Create/register store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStoreRequest"
      responses:
        "201":
          description: Store created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"

  /stores/{id}:
    get:
      operationId: getStore
      tags: [stores]
      summary: Get store by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Store details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"

  /analytics/summary:
    get:
      operationId: getAnalyticsSummary
      tags: [analytics]
      summary: Get analytics summary
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, 7days, 30days, 90days]
      responses:
        "200":
          description: Analytics summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsSummary"

  /analytics/chart:
    get:
      operationId: getAnalyticsChart
      tags: [analytics]
      summary: Get chart data
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7days, 30days, 90days]
      responses:
        "200":
          description: Chart data
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ChartDataPoint"
                required:
                  - data

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        price:
          type: number
        oldPrice:
          type: number
        description:
          type: string
        images:
          type: array
          items:
            type: string
        categoryId:
          type: string
        categoryName:
          type: string
        storeId:
          type: string
        storeName:
          type: string
        rating:
          type: number
        reviewCount:
          type: integer
        colors:
          type: array
          items:
            type: string
        sizes:
          type: array
          items:
            type: string
        dimensions:
          type: string
        warranty:
          type: string
        deliveryDays:
          type: integer
        isTopSelling:
          type: boolean
        isFeatured:
          type: boolean
        discount:
          type: integer
        salesCount:
          type: integer
      required:
        - id
        - name
        - price
        - storeId
        - storeName
        - categoryId

    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        icon:
          type: string
        productCount:
          type: integer
      required:
        - id
        - name

    Order:
      type: object
      properties:
        id:
          type: string
        orderNumber:
          type: string
        status:
          type: string
          enum: [new, accepted, shipped, delivered, cancelled]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        customerName:
          type: string
        customerPhone:
          type: string
        address:
          type: string
        comment:
          type: string
        paymentMethod:
          type: string
          enum: [cash, card, installment]
        totalPrice:
          type: number
        createdAt:
          type: string
        storeId:
          type: string
        storeName:
          type: string
      required:
        - id
        - orderNumber
        - status
        - items
        - totalPrice
        - createdAt

    OrderItem:
      type: object
      properties:
        productId:
          type: string
        productName:
          type: string
        productImage:
          type: string
        quantity:
          type: integer
        price:
          type: number
        color:
          type: string
      required:
        - productId
        - productName
        - quantity
        - price

    CreateOrderRequest:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: string
              quantity:
                type: integer
              color:
                type: string
            required:
              - productId
              - quantity
        customerName:
          type: string
        customerPhone:
          type: string
        address:
          type: string
        comment:
          type: string
        paymentMethod:
          type: string
          enum: [cash, card, installment]
      required:
        - items
        - customerName
        - customerPhone
        - address
        - paymentMethod

    Store:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        logo:
          type: string
        coverImage:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [official, manufacturer, partner]
        rating:
          type: number
        reviewCount:
          type: integer
        productCount:
          type: integer
        salesCount:
          type: integer
        deliveryRate:
          type: number
        location:
          type: string
        phone:
          type: string
        stir:
          type: string
        activityType:
          type: string
        isVerified:
          type: boolean
      required:
        - id
        - name
        - type

    CreateStoreRequest:
      type: object
      properties:
        name:
          type: string
        activityType:
          type: string
        stir:
          type: string
        location:
          type: string
        phone:
          type: string
        description:
          type: string
      required:
        - name
        - activityType
        - location
        - phone

    AnalyticsSummary:
      type: object
      properties:
        totalRevenue:
          type: number
        revenueChange:
          type: number
        totalOrders:
          type: integer
        ordersChange:
          type: number
        averageCheck:
          type: number
        averageCheckChange:
          type: number
        conversionRate:
          type: number
        conversionChange:
          type: number
        activeProducts:
          type: integer
        topProducts:
          type: array
          items:
            type: object
            properties:
              productId:
                type: string
              productName:
                type: string
              revenue:
                type: number
              salesCount:
                type: integer
            required:
              - productId
              - productName
              - revenue
              - salesCount
      required:
        - totalRevenue
        - revenueChange
        - totalOrders
        - ordersChange
        - averageCheck
        - averageCheckChange
        - conversionRate
        - conversionChange
        - activeProducts
        - topProducts

    ChartDataPoint:
      type: object
      properties:
        date:
          type: string
        revenue:
          type: number
        orders:
          type: integer
      required:
        - date
        - revenue
        - orders